# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Patoughi Driver API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/versions": {
            "get": {
                "tags": [
                    "Version"
                ],
                "summary": "Get Latest Version",
                "operationId": "Get Latest Version",
                "parameters": [
                    {
                        "name": "platform",
                        "in": "query",
                        "description": "Platform",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Type",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/VersionResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseErrorValidation"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error during registration",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseErrorServer"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/versions/check": {
            "get": {
                "tags": [
                    "Version"
                ],
                "summary": "Version",
                "operationId": "Version",
                "parameters": [
                    {
                        "name": "platform",
                        "in": "query",
                        "description": "Platform",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "version",
                        "in": "query",
                        "description": "version",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Type",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/VersionResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseErrorValidation"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error during registration",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseErrorServer"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "VersionResource": {
                "required": [
                    "platform",
                    "version",
                    "type",
                    "status",
                    "build_number",
                    "force_update"
                ],
                "properties": {
                    "platform": {
                        "type": "string",
                        "enum": [
                            "android",
                            "ios",
                            "pwa"
                        ]
                    },
                    "version": {
                        "type": "string"
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "driver",
                            "customer"
                        ]
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "draft",
                            "released",
                            "deprecated",
                            "inactive"
                        ]
                    },
                    "build_number": {
                        "type": "string"
                    },
                    "force_update": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "ApiResponseErrorValidation": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Server error message"
                    },
                    "errors": {
                        "description": "Validation errors",
                        "type": "object",
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object"
            },
            "ApiResponseErrorServer": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Server error message"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}